Validation Job exclude

+++ app/Controllers/Http/JobController.js
@@ -1,11 +1,11 @@
-'use strict'
-'use strict'
+"use strict";
+"use strict";
 
 /** @typedef {import('@adonisjs/framework/src/Request')} Request */
 /** @typedef {import('@adonisjs/framework/src/Response')} Response */
 /** @typedef {import('@adonisjs/framework/src/View')} View */
 
-const Job = use('App/Models/Job');
+const Job = use("App/Models/Job");
 
 /**
  * Resourceful controller for interacting with categories
@@ -20,11 +20,11 @@ class JobController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-   async index({ request, response, view }) {
+  async index({ request, response, view }) {
     const { page, qty, name } = request.all();
-    const query = Job.query().orderBy("name", "asc").with('categories');
-    if( name ) {
-      query.where('name', 'like', `%${name}%`).fetch();
+    const query = Job.query().orderBy("name", "asc").with("categories");
+    if (name) {
+      query.where("name", "like", `%${name}%`).fetch();
     }
     return await query.paginate(page, qty);
   }
@@ -36,7 +36,7 @@ class JobController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async store({ request, response }){
+  async store({ request, response }) {
     const registerFields = Job.getFields();
     const data = request.only(registerFields);
     return await Job.create(data);
@@ -52,14 +52,14 @@ class JobController {
    * @param {View} ctx.view
    */
   async show({ params, request, response, view }) {
-    
-    const data = await Job.query()
-      .where('id', params.id)
-      .first();
+    const data = await Job.query().where("id", params.id).first();
     data.hits = data.hits + 1;
     await data.save();
-    return await Job.query().where('id', params.id).with('users').with('categories').first();
-   
+    return await Job.query()
+      .where("id", params.id)
+      .with("users")
+      .with("categories")
+      .first();
   }
 
   /**
@@ -71,12 +71,12 @@ class JobController {
    * @param {Response} ctx.response
    */
   async update({ params, request, response }) {
-    const fields = Job.getFields()
-    const data = request.only(fields)
-    
-    const job = await Job.findOrFail(params.id)
-    
-    job.merge(data)
+    const fields = Job.getFields();
+    const data = request.only(fields);
+
+    const job = await Job.findOrFail(params.id);
+
+    job.merge(data);
     await job.save();
 
     return job;
@@ -90,16 +90,26 @@ class JobController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-   async destroy ({ params, request, response }) {
-    const job = await Job.findOrFail(params.id)
-
-    job.delete()
-
-    return {
-      message: 'Serviço removido com sucesso!'
-    };
+  async destroy({ params, request, response }) {
+    try{
+      const job = await Job.findOrFail(params.id);
+      if (job) {
+        response.status(200).json({
+          message: "Atividade removida com sucesso!",
+        });
+        job.delete();
+      } else {
+        response.status(404).json({
+          message: "Atividade não encontrada."
+        });
+      }
+    } catch(err){
+      response.status(401).json({
+        message: "Essa atividade já foi excluída ou não existe."
+      });
+    }
+    
+  }
 }
-}
 
-module.exports = JobController
-
+module.exports = JobController;
